pyproject.toml
[project]
name = "telegramtastic"
version = "0.1.0"
description = "A Meshtastic to thermal printer bridge"
authors = [
    {name = "Jon Davis", email = "github@konsoletek.com"},
]
dependencies = [
    "meshtastic",
    "paho-mqtt",
    "cryptography",
    "sqlalchemy",
    "python-dotenv",
    "pillow>=11.3.0",
    "python-escpos>=3.1",
]
requires-python = ">=3.13"
readme = "README.md"
license = {text = "MIT"}


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["common", "database"]

[tool.uv]
dev-dependencies = []

[project.urls]
Homepage = "https://github.com/shakataganai/telegramtastic"
Repository = "https://github.com/shakataganai/telegramtastic"
